separated communication functions into their own file
